+++ RELEASES.md
@@ -36,6 +36,8 @@ Upcoming
 - Support for new ESB standard (without ServiceContext)
 - Bugfix ignoreUnknownNamespaces default not true when noNamespaceSchemaLocation is being used
 - Don't add namespace to schema by default when targetNamespace present and default namespace is not
+- Determine default application server type (and remove version from type)
+
  
 
 ### Non backwards compatible changes